Fix all warinings from MBS building excluding const casting

+++ src/coap_options.c
@@ -36,7 +36,7 @@ uint16_t KNOWN_OPTIONS[ ] =
     OPT_NUM_URI_PORT,
     OPT_NUM_URI_HOST
 };
-size_t KNOWN_OPTIONS_COUNT = sizeof(KNOWN_OPTIONS) / sizeof(KNOWN_OPTIONS[0]);
+uint32_t KNOWN_OPTIONS_COUNT = sizeof(KNOWN_OPTIONS) / sizeof(KNOWN_OPTIONS[0]);
 
 //#########################################################################################################
 //### This function packs multiple CoAP options to the format specified at
